Extract `WaitForUnixSocket*` from `//cuttlefish/common/libs/utils:files.h`
This breaks a circular dependency between `files.h` and `subprocess.h`.

Bug: b/407763557
