+++ base/cvd/cuttlefish/common/libs/utils/files.h
@@ -107,9 +107,6 @@ Result<void> WalkDirectory(
 
 #ifdef __linux__
 Result<void> WaitForFile(const std::string& path, int timeoutSec);
-Result<void> WaitForUnixSocket(const std::string& path, int timeoutSec);
-Result<void> WaitForUnixSocketListeningWithoutConnect(const std::string& path,
-                                                      int timeoutSec);
 #endif
 
 // parameter to EmulateAbsolutePath
