+++ base/cvd/cuttlefish/host/commands/run_cvd/launch/vhost_device_vsock.cpp
@@ -28,10 +28,10 @@
 #include <fruit/fruit_forward_decls.h>
 #include <fruit/macro.h>
 
-#include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/known_paths.h"
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/wait_for_unix_socket.h"
 #include "cuttlefish/host/commands/run_cvd/launch/log_tee_creator.h"
 #include "cuttlefish/host/libs/config/command_source.h"
 #include "cuttlefish/host/libs/config/config_utils.h"
