+++ base/cvd/cuttlefish/host/commands/run_cvd/launch/wmediumd_server.cpp
@@ -24,9 +24,9 @@
 #include <fruit/fruit_forward_decls.h>
 #include <fruit/macro.h>
 
-#include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/wait_for_unix_socket.h"
 #include "cuttlefish/host/commands/run_cvd/launch/grpc_socket_creator.h"
 #include "cuttlefish/host/commands/run_cvd/launch/log_tee_creator.h"
 #include "cuttlefish/host/libs/config/command_source.h"
