+++ base/cvd/cuttlefish/host/libs/vm_manager/qemu_manager.cpp
@@ -39,6 +39,7 @@
 #include "common/libs/utils/files.h"
 #include "common/libs/utils/result.h"
 #include "common/libs/utils/subprocess.h"
+#include "common/libs/utils/wait_for_unix_socket.h"
 #include "host/libs/config/command_source.h"
 #include "host/libs/config/config_constants.h"
 #include "host/libs/config/cuttlefish_config.h"
